Update Inviscid_2D_Unconstrained_NACA0012.md
In line 165- It is mentioned that 'The first value in the parentheses is the variable type, which is 1 for a Hicks-Henne bump function'. But I believe the statement should be

'The first value in the parentheses is the variable type, which is 30 for a Hicks-Henne bump function'. 

Please check this. Thanks